Rename serviceHooks file
